rework validation so it works for both checkout and cart
