+++ assets/js/blocks/cart-checkout/checkout/style.scss
@@ -113,7 +113,7 @@
 			}
 
 			.wc-block-text-input,
-			.wc-block-select {
+			.wc-block-country-input {
 				float: left;
 				margin-left: #{$gap-small / 2};
 				margin-right: #{$gap-small / 2};
